rusk-wallet: Move interactive online and offline menu in separate files
