+++ rusk-wallet/src/bin/interactive.rs
@@ -4,23 +4,21 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+mod command_menu;
+
 use bip39::{Language, Mnemonic, MnemonicType};
 use requestty::Question;
 use rusk_wallet::{
     currency::Dusk,
     dat::{DatFileVersion, LATEST_VERSION},
-    gas::{self},
     Address, Error, Profile, Wallet, WalletPath, MAX_PROFILES,
 };
 
-use crate::io;
-use crate::io::prompt::request_auth;
-use crate::io::GraphQL;
-use crate::prompt;
-use crate::settings::Settings;
-use crate::Menu;
-use crate::WalletFile;
-use crate::{Command, RunResult};
+use crate::{
+    io::{self, prompt},
+    settings::Settings,
+    Command, GraphQL, Menu, RunResult, WalletFile,
+};
 
 /// Run the interactive UX loop with a loaded wallet
 pub(crate) async fn run_loop(
@@ -46,7 +44,7 @@ pub(crate) async fn run_loop(
                 // if the version file is old, ask for password and save as
                 // latest dat file
                 if file_version.is_old() {
-                    let pwd = request_auth(
+                    let pwd = prompt::request_auth(
                         "Updating your wallet data file, please enter your wallet password ",
                         password,
                         DatFileVersion::RuskBinaryFileFormat(LATEST_VERSION),
@@ -102,7 +100,7 @@ pub(crate) async fn run_loop(
 
             // request operation to perform
             let op = match wallet.is_online().await {
-                true => menu_op(
+                true => command_menu::online(
                     profile_idx,
                     wallet,
                     phoenix_spendable,
@@ -110,7 +108,7 @@ pub(crate) async fn run_loop(
                     settings,
                     is_synced,
                 ),
-                false => menu_op_offline(profile_idx, settings),
+                false => command_menu::offline(profile_idx, settings),
             };
 
             // perform operations with this profile
@@ -218,271 +216,6 @@ fn menu_profile(wallet: &Wallet<WalletFile>) -> anyhow::Result<ProfileSelect> {
 enum ProfileOp {
     Run(Box<Command>),
     Back,
-}
-
-#[derive(PartialEq, Eq, Hash, Clone, Debug)]
-enum CommandMenuItem {
-    // History
-    History,
-    // Transfer
-    Transfer,
-    // Stake
-    Stake,
-    // Unstake
-    Unstake,
-    // Withdraw
-    Withdraw,
-    // Contract Deploy
-    ContractDeploy,
-    // Contract Call
-    ContractCall,
-    // Conversion
-    Unshield,
-    Shield,
-    // Generate Contract ID.
-    CalculateContractId,
-    // Others
-    StakeInfo,
-    Export,
-    Back,
-}
-
-/// Allows the user to choose the operation to perform for the
-/// selected profile
-fn menu_op(
-    profile_idx: u8,
-    wallet: &Wallet<WalletFile>,
-    phoenix_balance: Dusk,
-    moonlight_balance: Dusk,
-    settings: &Settings,
-    is_synced: bool,
-) -> anyhow::Result<ProfileOp> {
-    use CommandMenuItem as CMI;
-
-    let mut cmd_menu = Menu::new()
-        .add(CMI::History, "Transactions History")
-        .add(CMI::Transfer, "Transfer")
-        .add(CMI::Unshield, "Convert shielded Dusk to public Dusk")
-        .add(CMI::Shield, "Convert public Dusk to shielded Dusk")
-        .add(CMI::StakeInfo, "Check Existing Stake")
-        .add(CMI::Stake, "Stake")
-        .add(CMI::Unstake, "Unstake")
-        .add(CMI::Withdraw, "Withdraw Stake Reward")
-        .add(CMI::StakeInfo, "Check Existing Stake")
-        .add(CMI::ContractDeploy, "Contract Deploy")
-        .add(CMI::ContractCall, "Contract Call")
-        .add(CMI::CalculateContractId, "Calculate Contract ID")
-        .add(CMI::Export, "Export provisioner key-pair")
-        .separator()
-        .add(CMI::Back, "Back")
-        .separator();
-
-    let msg = if !is_synced {
-        cmd_menu = Menu::new()
-            .add(CMI::Export, "Export provisioner key-pair")
-            .separator()
-            .add(CMI::Back, "Back")
-            .separator();
-
-        "The wallet is still syncing. Please come back to display new information."
-    } else {
-        "What do you like to do?"
-    };
-
-    let q = Question::select("theme")
-        .message(msg)
-        .choices(cmd_menu.clone())
-        .build();
-
-    let answer = requestty::prompt_one(q)?;
-    let cmd = cmd_menu.answer(&answer).to_owned();
-
-    let res = match cmd {
-        CMI::Transfer => {
-            let rcvr = prompt::request_rcvr_addr("recipient")?;
-            let (sender, balance) = match &rcvr {
-                Address::Shielded(_) => {
-                    (wallet.shielded_address(profile_idx)?, phoenix_balance)
-                }
-                Address::Public(_) => {
-                    (wallet.public_address(profile_idx)?, moonlight_balance)
-                }
-            };
-
-            let memo = Some(prompt::request_str("memo")?);
-            let amt = if memo.is_some() {
-                prompt::request_optional_token_amt("transfer", balance)
-            } else {
-                prompt::request_token_amt("transfer", balance)
-            }?;
-
-            ProfileOp::Run(Box::new(Command::Transfer {
-                sender: Some(sender),
-                rcvr,
-                amt,
-                gas_limit: prompt::request_gas_limit(
-                    gas::DEFAULT_LIMIT_TRANSFER,
-                )?,
-                memo,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-        CMI::Stake => {
-            let (addr, balance) = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    (wallet.shielded_address(profile_idx)?, phoenix_balance)
-                }
-                prompt::TransactionModel::Public => {
-                    (wallet.public_address(profile_idx)?, moonlight_balance)
-                }
-            };
-            ProfileOp::Run(Box::new(Command::Stake {
-                address: Some(addr),
-                amt: prompt::request_stake_token_amt(balance)?,
-                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-        CMI::Unstake => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
-            ProfileOp::Run(Box::new(Command::Unstake {
-                address: Some(addr),
-                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-
-        CMI::Withdraw => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
-            ProfileOp::Run(Box::new(Command::Withdraw {
-                address: Some(addr),
-                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-        CMI::ContractDeploy => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
-            ProfileOp::Run(Box::new(Command::ContractDeploy {
-                address: Some(addr),
-                code: prompt::request_contract_code()?,
-                init_args: prompt::request_bytes("init arguments")?,
-                deploy_nonce: prompt::request_nonce()?,
-                gas_limit: prompt::request_gas_limit(
-                    gas::DEFAULT_LIMIT_DEPLOYMENT,
-                )?,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-        CMI::ContractCall => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
-            ProfileOp::Run(Box::new(Command::ContractCall {
-                address: Some(addr),
-                contract_id: prompt::request_bytes("contract id")?,
-                fn_name: prompt::request_str("function name to call")?,
-                fn_args: prompt::request_bytes(
-                    "arguments of calling function",
-                )?,
-                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-                gas_price: prompt::request_gas_price()?,
-            }))
-        }
-        CMI::History => {
-            let profile_idx = Some(profile_idx);
-            ProfileOp::Run(Box::new(Command::History { profile_idx }))
-        }
-        CMI::StakeInfo => ProfileOp::Run(Box::new(Command::StakeInfo {
-            profile_idx: Some(profile_idx),
-            reward: false,
-        })),
-        CMI::Shield => ProfileOp::Run(Box::new(Command::Shield {
-            profile_idx: Some(profile_idx),
-            amt: prompt::request_token_amt("convert", moonlight_balance)?,
-            gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-            gas_price: prompt::request_gas_price()?,
-        })),
-        CMI::Unshield => ProfileOp::Run(Box::new(Command::Unshield {
-            profile_idx: Some(profile_idx),
-            amt: prompt::request_token_amt("convert", phoenix_balance)?,
-            gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-            gas_price: prompt::request_gas_price()?,
-        })),
-        CMI::CalculateContractId => {
-            ProfileOp::Run(Box::new(Command::CalculateContractId {
-                profile_idx: Some(profile_idx),
-                deploy_nonce: prompt::request_nonce()?,
-                code: prompt::request_contract_code()?,
-            }))
-        }
-        CMI::Export => ProfileOp::Run(Box::new(Command::Export {
-            profile_idx: Some(profile_idx),
-            name: None,
-            dir: prompt::request_dir("export keys", settings.profile.clone())?,
-        })),
-        CMI::Back => ProfileOp::Back,
-    };
-    Ok(res)
-}
-
-/// Allows the user to choose the operation to perform for the
-/// selected profile while in offline mode
-fn menu_op_offline(
-    profile_idx: u8,
-    settings: &Settings,
-) -> anyhow::Result<ProfileOp> {
-    use CommandMenuItem as CMI;
-
-    let cmd_menu = Menu::new()
-        .separator()
-        .add(CMI::Export, "Export provisioner key-pair")
-        .separator()
-        .add(CMI::Back, "Back");
-
-    let q = Question::select("theme")
-        .message("[OFFLINE] What would you like to do?")
-        .choices(cmd_menu.clone())
-        .build();
-
-    let answer = requestty::prompt_one(q)?;
-    let cmd = cmd_menu.answer(&answer).to_owned();
-
-    let res = match cmd {
-        CMI::Export => ProfileOp::Run(Box::new(Command::Export {
-            profile_idx: Some(profile_idx),
-            name: None,
-            dir: prompt::request_dir("export keys", settings.profile.clone())?,
-        })),
-        CMI::Back => ProfileOp::Back,
-        _ => unreachable!(),
-    };
-    Ok(res)
 }
 
 /// Allows the user to load a wallet interactively
